requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.6.0
langgraph-checkpoint>=2.0.0
chromadb>=0.4.22
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
faker>=21.0.0
pandas>=2.1.0
python-dateutil>=2.8.0
